\name{biovizBase-package}
\alias{biovizBase-package}
\alias{biovizBase-package}
\title{biovizBase is a package which provides utilities and color scheme...}
\description{biovizBase is a package which provides utilities and color scheme
for higher level graphic package which aim to visualize biological
data especially genetic data.}
\details{This package provides default color scheme for nucleotide, strand,
amino acid, try to pass colorblind checking as possible as we can.
And also provide giemsa stain result color scheme used to show
cytoband. This package also provides utilites to manipulate and
summarize raw data to get them ready to be visualized.}
\alias{biovizBase}
\docType{package}

