\name{CRC}
\alias{CRC}
\alias{crc.gr}
\alias{hg19sub}
\alias{mut.gr}
\title{CRC}
\usage{data(CRC)}
\description{Genomic sequencing of colorectal adenocarcinomas identifies a recurrent VTI1A-TCF7L2 fusion sample data set}
\details{The data used in
this study is from this a
paper http://www.nature.com/ng/journal/v43/n10/full/ng.936.html.}
\docType{data}
\keyword{datasets}
\examples{data(CRC)}