\name{crc1.GeRL}
\alias{crc1.GeRL}
\alias{crc1.GeRL}
\title{crc1.GeRL}
\usage{data(crc1.GeRL)}
\description{CRC-1 mutation and structural rearrangment}
\details{
  GenomicRangesList contains somatic mutation, reaarangment etc for a
  tumor sample, please check the reference for detials.
}
\docType{data}
\keyword{datasets}
\references{
  Genomic sequencing of colorectal adenocarcinomas identifies a recurrent VTI1A-TCF7L2 fusion}
\examples{data(crc1.GeRL)
crc1.GeRL
}

