\name{darned_hg19_subset500}
\alias{darned_hg19_subset500}
\alias{darned_hg19_subset500}
\title{Subset of RNA editing sites in hg19...}
\usage{data(darned_hg19_subset500)}
\description{Subset of RNA editing sites in hg19}
\details{This data set provides a subset(500 sites only) of hg19 RNA editing sites, and
originally from DARNED \url{http://darned.ucc.ie/} for the hg19
assembly.}
\docType{data}
\keyword{datasets}
\examples{data(darned_hg19_subset500)
darned_hg19_subset500}

