\name{genesymbol}
\alias{genesymbol}
\alias{genesymbol}
\title{Gene symbols with position...}
\usage{data(genesymbol)}
\description{Gene symbols with position}
\details{This data set provides genen symbols in human with position
and starnd information, stored as GRanges object.}
\docType{data}
\keyword{datasets}
\examples{
data(genesymbol)
head(genesymbol)
genesymbol["RBM17"]
}

