\name{hg19Ideogram}
\alias{hg19Ideogram}
\alias{hg19Ideogram}
\title{Hg19 ideogram without cytoband information...}
\usage{data(hg19Ideogram)}
\description{Hg19 ideogram without cytoband information}
\details{This data set provides hg19 genome information wihout cytoband
information.}
\docType{data}
\keyword{datasets}
\examples{data(hg19Ideogram)
hg19Ideogram}

