\name{hg19IdeogramCyto}
\alias{hg19IdeogramCyto}
\alias{hg19IdeogramCyto}
\title{Hg19 ideogram with cytoband information...}
\usage{data(hg19IdeogramCyto)}
\description{Hg19 ideogram with cytoband information}
\details{This data set provides hg19 genome information with cytoband
information.}
\docType{data}
\keyword{datasets}
\examples{data(hg19IdeogramCyto)
hg19IdeogramCyto}

