\name{ideo}
\alias{ideo}
\title{ideogram without cytoband information}
\usage{data(ideo)}
\description{ideogram without cytoband information}
\details{This data set provides hg19, hg18, mm10, mm9 genome information wihout cytoband
information as a lit.}
\docType{data}
\keyword{datasets}
\examples{data(ideo)
ideo}