\name{ideoCyto}
\alias{ideoCyto}
\title{ideogram with cytoband information}
\usage{data(ideoCyto)}
\description{ideogram with cytoband information}
\details{This data set provides hg19, hg18, mm10, mm9 genome information with cytoband
information as a lit.}
\docType{data}
\keyword{datasets}
\examples{data(ideoCyto)
ideoCyto}