\name{getYLab-method}
\alias{getYLab}
\alias{getYLab,TranscriptDb-method}
\title{parse y label information}
\description{
  parse y label information, object specific.
}
\usage{
\S4method{getYLab}{TranscriptDb}(obj)
}
\arguments{
  \item{obj}{
    A \code{TranscriptDb} object.
  }
}
\value{
  a string.
}
\author{Tengfei Yin}

